backend/requirements.txt
# DepthLens 后端依赖 (Phase 1: REST 基础骨架)
#
# 安装:
#   python -m venv venv
#   venv\Scripts\activate          (PowerShell: .\venv\Scripts\Activate.ps1)
#   pip install -r requirements.txt
#
# 注意:
# - torch / torchvision 在 Windows + CUDA 上推荐通过 PyTorch 官方 index 安装,
#   先跑下面这条:
#       pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   再 ``pip install -r requirements.txt`` 安装其余包.
# - 若仅 CPU 推理 (无独显), 用默认 PyPI 即可:
#       pip install torch torchvision
# - Phase 1 还没有 WebSocket 流, 所以暂不锁 ``websockets`` (FastAPI 自带 starlette
#   即可处理 WS, Phase 3 视情况再固定版本).

# ── Web 框架 ──────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12      # multipart/form-data 上传支持
pydantic==2.9.2

# ── 图像处理 ──────────────────────────────────────────────
Pillow==10.4.0
numpy==2.1.2
opencv-python==4.10.0.84       # Phase 3 视频管线会用, 提前装好

# ── 模型推理 (按需自行装 torch, 见上方说明) ─────────────
transformers==4.46.0
# torch>=2.2,<3                 # 自行通过 pytorch.org 索引安装
# torchvision>=0.17,<1

# ── 系统信息 ──────────────────────────────────────────────
psutil==6.1.0

# ── 测试 (Phase 1 验证用) ─────────────────────────────────
pytest==8.3.3
httpx==0.27.2                  # FastAPI TestClient 依赖
